회원 검색 (지역)

+++ src/main/java/com/sinsaflower/server/domain/member/dto/MemberProductPriceResponse.java
@@ -3,11 +3,13 @@ package com.sinsaflower.server.domain.member.dto;
 import lombok.Builder;
 import lombok.Getter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Builder
 public class MemberProductPriceResponse {
 
     private String categoryName;
-    private Integer price;       // 천원 단위
+    private BigDecimal price;       // 천원 단위
     private Boolean isAvailable;
 }
